+++ TeacherClient/MainForm.Designer.cs
@@ -111,6 +111,8 @@ partial class MainForm
         var filesMenuItem = new ToolStripMenuItem(TeacherClientText.FilesMenu);
         filesMenuItem.DropDownItems.Add(TeacherClientText.RefreshBoth, null, refreshFilesButton_Click);
         filesMenuItem.DropDownItems.Add($"{TeacherClientText.Upload} ->", null, uploadButton_Click);
+        filesMenuItem.DropDownItems.Add(TeacherClientText.SendToSelectedStudents, null, sendToSelectedStudentsButton_Click);
+        filesMenuItem.DropDownItems.Add(TeacherClientText.SendToAllOnlineStudents, null, sendToAllOnlineStudentsButton_Click);
         filesMenuItem.DropDownItems.Add($"<- {TeacherClientText.Download}", null, downloadButton_Click);
         filesMenuItem.DropDownItems.Add(TeacherClientText.DeleteLocal, null, deleteLocalButton_Click);
         filesMenuItem.DropDownItems.Add(TeacherClientText.DeleteRemote, null, deleteRemoteButton_Click);
@@ -200,6 +202,7 @@ partial class MainForm
         ConfigureGrid(processesGrid);
         ConfigureGrid(localFilesGrid);
         ConfigureGrid(remoteFilesGrid);
+        agentsGrid.MultiSelect = true;
 
         agentsGrid.Dock = DockStyle.Fill;
         agentsGrid.CellDoubleClick += agentsGrid_CellDoubleClick;
@@ -377,6 +380,8 @@ partial class MainForm
         filesToolStrip.Items.Add(CreateToolbarButton(TeacherClientText.RefreshBoth, ToolbarIconKind.Refresh, refreshFilesButton_Click));
         filesToolStrip.Items.Add(new ToolStripSeparator());
         filesToolStrip.Items.Add(CreateToolbarButton(TeacherClientText.Upload, ToolbarIconKind.Upload, uploadButton_Click));
+        filesToolStrip.Items.Add(CreateToolbarButton(TeacherClientText.SendToSelectedStudents, ToolbarIconKind.UploadGroup, sendToSelectedStudentsButton_Click));
+        filesToolStrip.Items.Add(CreateToolbarButton(TeacherClientText.SendToAllOnlineStudents, ToolbarIconKind.Broadcast, sendToAllOnlineStudentsButton_Click));
         filesToolStrip.Items.Add(CreateToolbarButton(TeacherClientText.Download, ToolbarIconKind.Download, downloadButton_Click));
         filesToolStrip.Items.Add(new ToolStripSeparator());
         filesToolStrip.Items.Add(CreateToolbarButton(TeacherClientText.DeleteLocal, ToolbarIconKind.Remove, deleteLocalButton_Click));
@@ -548,7 +553,9 @@ partial class MainForm
             ToolbarIconKind.Stop => Color.FromArgb(190, 24, 93),
             ToolbarIconKind.Folder => Color.FromArgb(217, 119, 6),
             ToolbarIconKind.Upload => Color.FromArgb(2, 132, 199),
+            ToolbarIconKind.UploadGroup => Color.FromArgb(37, 99, 235),
             ToolbarIconKind.Download => Color.FromArgb(14, 116, 144),
+            ToolbarIconKind.Broadcast => Color.FromArgb(124, 58, 237),
             ToolbarIconKind.NewFolder => Color.FromArgb(202, 138, 4),
             _ => Color.FromArgb(71, 85, 105)
         };
@@ -617,11 +624,25 @@ partial class MainForm
                 graphics.DrawLine(pen, 18, 10, 14, 6);
                 graphics.DrawLine(pen, 7, 22, 21, 22);
                 break;
+            case ToolbarIconKind.UploadGroup:
+                graphics.DrawLine(pen, 10, 18, 10, 7);
+                graphics.DrawLine(pen, 7, 10, 10, 7);
+                graphics.DrawLine(pen, 13, 10, 10, 7);
+                graphics.DrawEllipse(pen, 4, 18, 5, 5);
+                graphics.DrawEllipse(pen, 12, 18, 5, 5);
+                graphics.DrawEllipse(pen, 20, 18, 5, 5);
+                break;
             case ToolbarIconKind.Download:
                 graphics.DrawLine(pen, 14, 6, 14, 22);
                 graphics.DrawLine(pen, 10, 18, 14, 22);
                 graphics.DrawLine(pen, 18, 18, 14, 22);
                 graphics.DrawLine(pen, 7, 6, 21, 6);
+                break;
+            case ToolbarIconKind.Broadcast:
+                graphics.FillEllipse(accentBrush, 12, 12, 4, 4);
+                graphics.DrawArc(pen, 8, 8, 12, 12, 315, 90);
+                graphics.DrawArc(pen, 5, 5, 18, 18, 315, 90);
+                graphics.DrawArc(pen, 2, 2, 24, 24, 315, 90);
                 break;
             case ToolbarIconKind.NewFolder:
                 graphics.DrawRectangle(pen, 4, 11, 14, 10);
@@ -648,7 +669,9 @@ partial class MainForm
         Stop,
         Folder,
         Upload,
+        UploadGroup,
         Download,
+        Broadcast,
         NewFolder
     }